cloud-functions/twitter-fetcher/requirements.txt
functions-framework==3.4.0
google-cloud-storage==2.10.0
google-cloud-secret-manager==2.16.4
requests==2.31.0
flask==2.3.3
